Mark BaseAI/Fighter as Actor only components.
I'm surprised it worked this quickly, I thought this would be harder
to do.

There's a small chance this actually is invalid, but for now MyPy will
verify the types the way I want it to.

This also fixes a type error that BaseAI had after the change to
BaseComponent.